Add files via upload
This file contains code that can perform all the tasks except for 4 (promoter detection) and 6 (5-mer counting) (can someone implement them?)
The main subroutine takes a sequence and do all the jobs. The integration of the sequence module (analysis.pl) and this file also needs to be done.

Math::Cartesian::Product is required to run the code. Use cpan to install it.